feat(BaseComponent): Rename Base/index into Base/BaseComponent
